Added a short tutorial to the README introduction.
This updates the example in the README to use the new feathers-mongoose syntax and provides a few jquery examples to better show how it works.  It'll save some people a few minutes by giving an example of how to use jQuery.ajax to send the custom Authorization header in the request.