add features and fix typos (#309)
* add features and fix typos

changed the user param to authResult which is more accurate.
Added a user variable from connection to allow the example to function in its entirety.
Added a check to "Array.IsArray" to the user.rooms to ensure it both exists and is an array.
Added another example of channel sorting for specified users via email and id
Placed a NOT RECOMMENDED comment after the app.publish for all authenticated users
Added a new publish method to demonstrate how a messages service would utilize the emails and ids to send to users involved.

* Update channels.js

* Minor change to trigger new build