use the old version of events.ts
replace the code from https://deno.land/x/events with the old adapted code from nodejs
This was also used before the updating of the packages for deno.
The problem with the code from the events library was, that the exported EventEmitter is a class and assigning the prototype of this class, as happens in the feathers package for the services, doesn't work and would cause many errors. 
I decided to go with a .js file because a function doesn't work as constructor in typescript and assigning to the prototype of classes is (afaik) not possible without ignoring the line for typescript