port transport-commons package
there is one problem: How can we tell if we are in production mode in Deno? 
In Node it was done via the env variable NODE_ENV, but in deno, this, of course, doesn't exist.
Using a Env variable is maybe not the best way to do it in deno, because then the `--allow-env` flag would be needed to run the code