fix(cli): mongodb connection string for node 17+
On Node version 17+, `localhost` gets converted to the IPV6 version `:::1`.  The MongoDB database doesn’t receive the connection.

This PR updates the default MongoDB connection string to use the IPV4 loopback address `127.0.0.1`, which restores the ability to connect to the MongoDB server.

One place in the docs was also updated.