Expose parsed files to hooks and services
When using the koa-body body parser it will create an additional property (files) in the request object. This property is not exposed to feather services / hooks which make it tedious to work with file uploads directly in the services itself.

This field will only be filled when multipart is activated and a binary file is sent. Otherwise it will be undefined.